Updated Fraction, Added Adaptive Runge-Kutta-Cash-Karp, modified simulator related things tied to frequency for updates
Updated fraction to be able to be initialized by a double decimal input.

Created an adaptive Butcher tableau framework and created a Runge-Kutta
Cash-Karp method to test it. Proved to work as expected on all the test
problems excluding the 3-Body problem. For some reason, the adaptive
method would produce identical estimates of the solution at the next
time, in turn making the relative error equal to zero. This would then
make it appear the step size was small enough to work. However, the
result doesn’t end up correct. Thoughts are this may be a problem that
occurs for certain Hamiltonian dynamical systems, but unsure. Tried
looking for any bugs but can’t find anything that should be, especially
after re-implementing the adaptive method in matlab and validating it
produces the same results.