Made progress in development of new sim framework
Have made a lot of progress getting the new framework put together.
Sill have work to do for the sim_base and figuring out the best way to
allow for it to generate two simulation versions.

Note that one simulation version will match what was done before, which
is having a integration scheme chosen at compile time and applying it
to all the models in the simulation. The other approach is making each
dynamic model choose what integration scheme it wants to use and then
integration with that scheme separately from the other methods. This
latter approach should allow for simulations that can add in new models
on the fly and allow for treating different dynamics uniquely,
hopefully generating better results.