Added to the Sim Base and built a Uniform Sim class, Added to History Printing Class and FileWrap class
Largely completed the base sim class, excluding functionality that
needs to be written to do a given time step increment. This part may
need some CRTP functions to handle doing a uniform integration versus
integrating individual models (the block approach) independently.

Built a uniform sim class that essentially requires an Integrator class
to be passed as a parameter.

Added to the history printer such that it can add an empty line, which
is used to add a line after a given monte carlo run when about to start
a new monte carlo run.

Made the FileWrap class capable of closing a file, which is something
that should have been implemented anyway.